Title: 합과 차의 나눗셈, Time: 101ms, Memory: 9MB, Status: Wrong Answer - CodeTree

+++ 240807/합과 차의 나눗셈/divide-of-sum-and-sub.java
@@ -7,7 +7,7 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         String[] input = br.readLine().split(" ");
         int num1 = Integer.parseInt(input[0]);
-        double num2 = Float.parseFloat(input[1]);
+        double num2 = Double.parseDouble(input[1]);
         double result = (num1 + num2) / (num1 - num2);
         result=Math.round(result*100) / 100.0;
         bw.write(Double.toString(result));
